requirements.txt
Django==2.1.3
djangorestframework==3.9.0
django-filter==2.0.0
django-rest-swagger==2.2.0
django-cors-headers==2.4.0
gunicorn==19.9.0
mysqlclient==1.3.13
drf-flex-fields==0.3.4